:truck: Refs #276 - Move files to a more central location
