--- src/openforms/appointments/contrib/jcc/tests/utils.py
@@ -1,9 +1,0 @@
-import os
-
-
-def mock_response(filename):
-    filepath = os.path.abspath(
-        os.path.join(os.path.dirname(__file__), "mock", filename)
-    )
-    with open(filepath, "r") as f:
-        return f.read()
